racha_app/requirements.txt
Flask
Flask-Migrate
Flask-SQLAlchemy
PyJWT
psycopg2-binary
celery
passlib
pandas
requests